refactor test explorer: move out test node
